팔로우 기능 구현

+++ src/main/java/com/cos/photogramstart/domain/image/ImageRepository.java
@@ -2,5 +2,5 @@ package com.cos.photogramstart.domain.image;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageRepository extends JpaRepository<Image, Long> {
+public interface ImageRepository extends JpaRepository<Image, Integer> {
 }
